Do not track resource downloads with incomplete info
For: https://github.com/openstax/os-webview/issues/2482
Also only use the PDF download dialog if the download is a PDF.
